% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floodplain_vol_table.R
\name{floodplain_vol_table}
\alias{floodplain_vol_table}
\title{Floodplain Volume Table}
\usage{
floodplain_vol_table(channel_vol, floodplain_vol)
}
\arguments{
\item{channel_vol}{numeric; Channel volume.}

\item{floodplain_vol}{numeric; Floodplain volume}
}
\value{
a gt table
}
\description{
Create a floodplain volume table.
}
