% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_message.R
\name{log_message}
\alias{log_message}
\title{Log a message}
\usage{
log_message(msg)
}
\arguments{
\item{msg}{character; The text of the message to display in the log.}
}
\value{
message condition
}
\description{
Logs a system message condition suitable for display in a 
Shiny console window.
}
