% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs_discharge_table.R
\name{xs_discharge_table}
\alias{xs_discharge_table}
\title{Create a Cross Section Discharge Table}
\usage{
xs_discharge_table(xs_pts, xs_number, bf_estimate, mannings_n)
}
\arguments{
\item{xs_pts}{sf; A cross section lines feature class.}

\item{xs_number}{integer; The cross section `Seq` number of the
requested cross section.}

\item{bf_estimate}{numeric; Detrended bankfull estimate (units:
detrended feet).}

\item{mannings_n}{numeric; The Manning's n coeficient.}
}
\value{
a `gt` object
}
\description{
Creates a cross section discharge table for the channel portion of the
specified cross section.
}
